+++ bindings/nodejs/lib/bindings.ts
@@ -27,6 +27,7 @@ const {
     getClientFromWallet,
     getSecretManagerFromWallet,
     migrateStrongholdSnapshotV2ToV3,
+    migrateDbFromChrysalisToStardust,
 } = addon;
 
 const callClientMethodAsync = (
@@ -116,4 +117,5 @@ export {
     getSecretManagerFromWallet,
     listenMqtt,
     migrateStrongholdSnapshotV2ToV3,
+    migrateDbFromChrysalisToStardust
 };