+++ sdk/src/wallet/storage/manager.rs
@@ -105,6 +105,13 @@ impl StorageManager {
         self.set(&key, &sync_options).await
     }
 
+    pub(crate) async fn get_chrysalis_data(
+        &self,
+    ) -> crate::wallet::Result<Option<std::collections::HashMap<String, String>>> {
+        self.get(crate::wallet::migration::chrysalis::CHRYSALIS_STORAGE_KEY)
+            .await
+    }
+
     pub(crate) async fn get_default_sync_options(
         &self,
         account_index: u32,