+++ bindings/nodejs/CHANGELOG.md
@@ -21,6 +21,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## 1.0.6 - 2023-08-25
 
+### Added
+
+- `migrateDbFromChrysalisToStardust` function;
+
 ### Fixed
 
 - `Account::prepareBurn()` return type;